requirements.txt
pyyaml>=6.0
qdrant-client>=1.7.0
requests>=2.31.0
rich>=13.0.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0